update product updates header
